# Test_external/Back/Swagger/openapi.json
{"openapi":"3.1.0","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/api/hydroponic/all":{"get":{"tags":["hydroponic"],"summary":"Get All Hydroponics","operationId":"get_all_hydroponics_api_hydroponic_all_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/Hydroponic_response"},"type":"array","title":"Response Get All Hydroponics Api Hydroponic All Get"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/hydroponic/{hydroponic_id}":{"get":{"tags":["hydroponic"],"summary":"Get Hydroponic By Id","operationId":"get_hydroponic_by_id_api_hydroponic__hydroponic_id__get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"hydroponic_id","in":"path","required":true,"schema":{"type":"integer","title":"Hydroponic Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"anyOf":[{"$ref":"#/components/schemas/Hydroponic_response"},{"type":"null"}],"title":"Response Get Hydroponic By Id Api Hydroponic  Hydroponic Id  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["hydroponic"],"summary":"Delete Hydroponic","operationId":"delete_hydroponic_api_hydroponic__hydroponic_id__delete","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"hydroponic_id","in":"path","required":true,"schema":{"type":"integer","title":"Hydroponic Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/hydroponic/add":{"post":{"tags":["hydroponic"],"summary":"Add Hydroponic","operationId":"add_hydroponic_api_hydroponic_add_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Hydroponic_input"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/hydroponic/reset/{hydroponic_id}":{"patch":{"tags":["hydroponic"],"summary":"Reset Hydroponic","operationId":"reset_hydroponic_api_hydroponic_reset__hydroponic_id__patch","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"hydroponic_id","in":"path","required":true,"schema":{"type":"integer","title":"Hydroponic Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/hydroponic/{hydroponic_id}/update/water/add_10_percent":{"patch":{"tags":["hydroponic"],"summary":"Add 10 Percent Water","operationId":"add_10_percent_water_api_hydroponic__hydroponic_id__update_water_add_10_percent_patch","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"hydroponic_id","in":"path","required":true,"schema":{"type":"integer","title":"Hydroponic Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/hydroponic/{hydroponic_id}/update/minerals/add_5_percent":{"patch":{"tags":["hydroponic"],"summary":"Add 5 Percent Minerals","operationId":"add_5_percent_minerals_api_hydroponic__hydroponic_id__update_minerals_add_5_percent_patch","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"hydroponic_id","in":"path","required":true,"schema":{"type":"integer","title":"Hydroponic Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/hydroponic/{hydroponic_id}/update/temperature/add_1_celsius":{"patch":{"tags":["hydroponic"],"summary":"Add 1 Celsius Temperature","operationId":"add_1_celsius_temperature_api_hydroponic__hydroponic_id__update_temperature_add_1_celsius_patch","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"hydroponic_id","in":"path","required":true,"schema":{"type":"integer","title":"Hydroponic Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/hydroponic/{hydroponic_id}/update/temperature/lower_1_celsius":{"patch":{"tags":["hydroponic"],"summary":"Lower 1 Celsius Temperature","operationId":"lower_1_celsius_temperature_api_hydroponic__hydroponic_id__update_temperature_lower_1_celsius_patch","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"hydroponic_id","in":"path","required":true,"schema":{"type":"integer","title":"Hydroponic Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/hydroponic/{hydroponic_id}/update/acidity/add_0_25":{"patch":{"tags":["hydroponic"],"summary":"Add 0 25 Acid","operationId":"add_0_25_acid_api_hydroponic__hydroponic_id__update_acidity_add_0_25_patch","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"hydroponic_id","in":"path","required":true,"schema":{"type":"integer","title":"Hydroponic Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/hydroponic/{hydroponic_id}/update/acidity/lower_0_25":{"patch":{"tags":["hydroponic"],"summary":"Lower 0 25 Acid","operationId":"lower_0_25_acid_api_hydroponic__hydroponic_id__update_acidity_lower_0_25_patch","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"hydroponic_id","in":"path","required":true,"schema":{"type":"integer","title":"Hydroponic Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/hydroponic/{hydroponic_id}/update/oxygen/add_5_percent":{"patch":{"tags":["hydroponic"],"summary":"Add 5 Percent Oxygen","operationId":"add_5_percent_oxygen_api_hydroponic__hydroponic_id__update_oxygen_add_5_percent_patch","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"hydroponic_id","in":"path","required":true,"schema":{"type":"integer","title":"Hydroponic Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/token":{"post":{"summary":"Login","operationId":"login_token_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_login_token_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Token"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/register":{"post":{"summary":"Register","operationId":"register_register_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_register_register_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Token"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/user/exists":{"get":{"summary":"User Exist","operationId":"user_exist_user_exists_get","parameters":[{"name":"username","in":"query","required":true,"schema":{"type":"string","minLength":1,"maxLength":120,"pattern":"^[a-zA-Z0-9_\\.]+$","title":"Username"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"boolean","title":"Response User Exist User Exists Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/user":{"delete":{"summary":"Delete User","operationId":"delete_user_user_delete","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/users/me":{"get":{"summary":"Read Users Me","operationId":"read_users_me_users_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/get_startup_time":{"get":{"summary":"Get Startup Time","operationId":"get_startup_time_get_startup_time_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"integer","title":"Response Get Startup Time Get Startup Time Get"}}}}}}}},"components":{"schemas":{"Body_login_token_post":{"properties":{"grant_type":{"anyOf":[{"type":"string","pattern":"password"},{"type":"null"}],"title":"Grant Type"},"username":{"type":"string","title":"Username"},"password":{"type":"string","title":"Password"},"scope":{"type":"string","title":"Scope","default":""},"client_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Id"},"client_secret":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Secret"}},"type":"object","required":["username","password"],"title":"Body_login_token_post"},"Body_register_register_post":{"properties":{"username":{"type":"string","maxLength":120,"minLength":1,"pattern":"^[a-zA-Z0-9_\\.]+$","title":"Username"},"password":{"type":"string","maxLength":120,"minLength":1,"pattern":"^[a-zA-Z0-9!@#$%\\^&*\\(\\)\\.,]+$","title":"Password"}},"type":"object","required":["username","password"],"title":"Body_register_register_post"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"Hydroponic_input":{"properties":{"name":{"type":"string","maxLength":255,"minLength":1,"title":"Name"},"water_amount":{"type":"number","exclusiveMinimum":0.0,"title":"Water Amount"},"water_consumption":{"type":"number","exclusiveMinimum":0.0,"title":"Water Consumption"},"minerals_amount":{"type":"number","exclusiveMinimum":0.0,"title":"Minerals Amount"},"minerals_optimal":{"type":"number","exclusiveMinimum":0.0,"title":"Minerals Optimal"},"minerals_consumption":{"type":"number","exclusiveMinimum":0.0,"title":"Minerals Consumption"},"acidity_optimal_ph":{"type":"number","maximum":14.0,"minimum":0.0,"title":"Acidity Optimal Ph"},"temperature_C_optimal":{"type":"number","maximum":25.0,"minimum":5.0,"title":"Temperature C Optimal"},"oxygen_amount":{"type":"number","exclusiveMinimum":0.0,"title":"Oxygen Amount"},"oxygen_consumption":{"type":"number","exclusiveMinimum":0.0,"title":"Oxygen Consumption"}},"type":"object","required":["name","water_amount","water_consumption","minerals_amount","minerals_optimal","minerals_consumption","acidity_optimal_ph","temperature_C_optimal","oxygen_amount","oxygen_consumption"],"title":"Hydroponic_input"},"Hydroponic_response":{"properties":{"id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Id"},"name":{"type":"string","title":"Name"},"water_amount":{"type":"number","title":"Water Amount"},"water_consumption":{"type":"number","title":"Water Consumption"},"minerals_amount":{"type":"number","title":"Minerals Amount"},"minerals_optimal":{"type":"number","title":"Minerals Optimal"},"minerals_consumption":{"type":"number","title":"Minerals Consumption"},"acidity_optimal_ph":{"type":"number","title":"Acidity Optimal Ph"},"temperature_C_optimal":{"type":"number","title":"Temperature C Optimal"},"oxygen_amount":{"type":"number","title":"Oxygen Amount"},"oxygen_consumption":{"type":"number","title":"Oxygen Consumption"},"value_water":{"type":"number","title":"Value Water"},"value_minerals":{"type":"number","title":"Value Minerals"},"value_acidity_ph":{"type":"number","title":"Value Acidity Ph"},"value_temperature_C":{"type":"number","title":"Value Temperature C"},"value_oxygen":{"type":"number","title":"Value Oxygen"}},"type":"object","required":["id","name","water_amount","water_consumption","minerals_amount","minerals_optimal","minerals_consumption","acidity_optimal_ph","temperature_C_optimal","oxygen_amount","oxygen_consumption","value_water","value_minerals","value_acidity_ph","value_temperature_C","value_oxygen"],"title":"Hydroponic_response"},"Token":{"properties":{"access_token":{"type":"string","title":"Access Token"},"token_type":{"type":"string","title":"Token Type"}},"type":"object","required":["access_token","token_type"],"title":"Token"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"OAuth2PasswordBearer":{"type":"oauth2","flows":{"password":{"scopes":{},"tokenUrl":"token"}}}}}}